Allows playback through gap > maxBufferMs
This change fixes #8959

The issue is caused because the current logic assumes it is fully buffered after
fetching `minBufferMs` worth of GAP segments, which of course contain no
samples.  ExoPlayer normally jumps past the GAP but the infinite buffering
stall prevents this.

The fix changes the HLS implementation of `SequenceableLoader.getBufferedPositionUs()`
so to not report a GAP segment as potentially adding to the buffered time.
This allows deferring to the `SampleQueue`, a more complete source of truth.

Note the jump to end of the GAP occurs with no buffer, and loading has not triggered
so at least one handler loop is required to prevent false trigger of the
"Playback stuck buffering and not loading" error.